pyproject.toml
[tool.poetry]
name = "inspector"
version = "0.2.0"
description = ""
authors = ["André Rocha"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
openai = "^1.54.4"
numpy = "^2.1.3"
langchain-text-splitters = "^0.3.2"
pytest = "^8.3.3"
tiktoken = "^0.8.0"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
streamlit = "^1.40.1"
streamlit-option-menu = "^0.4.0"
redis = "^5.2.0"
redisvl = "^0.3.6"
tika = "^2.6.0"
aiohttp = "^3.11.2"
python-multipart = "^0.0.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
